refactor: flatten header components under layout
