+++ src/components/layout/DesktopNav.tsx
@@ -7,7 +7,7 @@ import * as React from 'react';
 import { UnderlineLink } from '@/components/ui';
 
 import { ServicesDropdownMenu } from './ServicesDropdownMenu';
-import type { NavItemProps } from './types';
+import type { NavItemProps } from './header.types';
 
 type DesktopNavProps = {
   items: NavItemProps[];
