+++ src/components/layout/Layout.tsx
@@ -5,7 +5,7 @@ import * as React from 'react';
 import { CallToAction } from '@/components/templates/CallToAction';
 
 import { Footer } from './Footer';
-import { Header } from './header/Header';
+import { Header } from './Header';
 
 export function Layout({ children }: { children: React.ReactNode }) {
   return (
