+++ src/components/layout/index.ts
@@ -1,8 +1,8 @@
 export * from './Container';
 export * from './Footer';
 export * from './InfoBar';
-export { Header } from './header/Header';
-export type { NavItemProps, SubItem } from './header/types';
+export * from './Header';
+export type { NavItemProps, SubItem } from './header.types';
 
 export { Layout } from './Layout';
 export { Page } from './Page';
